Gauntlet > Gauntlet Extended
Rename Gauntlet to Gauntlet Extended
