+++ gauntletextended/src/main/java/net/runelite/client/plugins/gauntletextended/overlay/OverlayHunllef.java
@@ -25,7 +25,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package net.runelite.client.plugins.gauntlet.overlay;
+package net.runelite.client.plugins.gauntletextended.overlay;
 
 import java.awt.Color;
 import java.awt.Dimension;
@@ -48,11 +48,11 @@ import net.runelite.api.Projectile;
 import net.runelite.api.coords.LocalPoint;
 import net.runelite.api.model.Jarvis;
 import net.runelite.api.model.Vertex;
-import net.runelite.client.plugins.gauntlet.GauntletConfig;
-import net.runelite.client.plugins.gauntlet.GauntletPlugin;
-import net.runelite.client.plugins.gauntlet.entity.Hunllef;
-import net.runelite.client.plugins.gauntlet.entity.Missile;
-import net.runelite.client.plugins.gauntlet.entity.Tornado;
+import net.runelite.client.plugins.gauntletextended.GauntletExtendedConfig;
+import net.runelite.client.plugins.gauntletextended.GauntletExtendedPlugin;
+import net.runelite.client.plugins.gauntletextended.entity.Hunllef;
+import net.runelite.client.plugins.gauntletextended.entity.Missile;
+import net.runelite.client.plugins.gauntletextended.entity.Tornado;
 import net.runelite.client.ui.overlay.OverlayLayer;
 import net.runelite.client.ui.overlay.OverlayPosition;
 import net.runelite.client.ui.overlay.OverlayPriority;
@@ -78,8 +78,8 @@ public class OverlayHunllef extends Overlay
 	private static final int COLOR_DURATION = 10;
 
 	private final Client client;
-	private final GauntletPlugin plugin;
-	private final GauntletConfig config;
+	private final GauntletExtendedPlugin plugin;
+	private final GauntletExtendedConfig config;
 	private final ModelOutlineRenderer modelOutlineRenderer;
 
 	private Hunllef hunllef;
@@ -88,7 +88,7 @@ public class OverlayHunllef extends Overlay
 	private int idx;
 
 	@Inject
-	private OverlayHunllef(final Client client, final GauntletPlugin plugin, final GauntletConfig config, final ModelOutlineRenderer modelOutlineRenderer)
+	private OverlayHunllef(final Client client, final GauntletExtendedPlugin plugin, final GauntletExtendedConfig config, final ModelOutlineRenderer modelOutlineRenderer)
 	{
 		super(plugin);
